+++ tox.ini
@@ -214,7 +214,7 @@ commands = python -m pip list --format=columns
 [testenv:snowpark_pandas_modin_pandas_import_error]
 description = test error messages when importing unsupported modin or pandas versions
 basepython = python3.9
-commands = python -m pytest --noconftest tests/unit/modin/test_python_version.py
+commands = TEST_INCORRECT_MODIN_PANDAS_VERSIONS=True python -m pytest --noconftest tests/unit/modin/test_modin_pandas_version.py
 
 [testenv:snowpark_pandas_old_np]
 description = test that Snowpark pandas works with numpy 1.26
